+++ CHANGELOG.md
@@ -165,3 +165,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - documentation typos
+
+## [0.12.0] - 2019-12-31
+
+- useGeolocationEvents hook & tests
+- useGeolocationState hook & tests
+- useGeolocation hook & tests